WIP: fix test with authorization header
