+++ ToDoApp/Backend/src/resolvers/todo/todos.spec.js
@@ -1,53 +1,18 @@
-const {
-    mergeResolvers
-} = require("merge-graphql-schemas");
-const {
-    createTestClient
-} = require('apollo-server-testing');
-const {
-    ApolloServer,
-    gql,
-    makeExecutableSchema
-} = require('apollo-server');
-const {
-    typeDefs
-} = require("../../schema/typeDefs");
-const {
-    userResolver
-} = require("../user/userResolver");
-const {
-    todoResolver
-} = require("./todoResolver");
-const {
-    augmentSchema
-} = require("neo4j-graphql-js");
-const neo4j = require('neo4j-driver');
+const { createTestClient } = require('apollo-server-testing');
+const { gql} = require('apollo-server');
+const { getTestApolloServer } = require('../../utils/testHelper');
 
-const driver = neo4j.driver(
-    'bolt://localhost',
-    neo4j.auth.basic('neo4j', 'password'), {
-        disableLosslessIntegers: true
-    }
-)
-const resolvers = mergeResolvers([userResolver, todoResolver]);
-
-const schema = makeExecutableSchema({
-    typeDefs,
-    resolvers
-});
-const augmentedSchema = augmentSchema(schema);
-
-const server = new ApolloServer({
-    schema: augmentedSchema,
-    context: {
-        driver
-    }
-});
+/*const {
+    query,
+    mutate,
+} = createTestClient(getTestApolloServer(true));*/
 
 const {
     query,
     mutate
-} = createTestClient(server);
+} = createTestClient(getTestApolloServer(true));
+
+
 /**
 describe('query', () => {
     describe('todos', () => {
@@ -75,8 +40,7 @@ describe('mutate', () => {
             const res = await mutate({
                 mutation: CREATE_TODO,
                 variables: {
-                    message: "neues Todo",
-                    token: ""
+                    message: "neues Todo"
                 }
             });
             expect(res.errors).toHaveLength(1);
@@ -87,8 +51,7 @@ describe('mutate', () => {
             const res = await mutate({
                 mutation: CREATE_TODO,
                 variables: {
-                    message: "neues Todo",
-                    token: logInToken
+                    message: "neues Todo"
                 }
             });
             expect(res.data.addTodo[0].message).toEqual("neues Todo");
